LTI: improve creation for content item types - refs BT#16379
